Fix BeforeAll hook
